#38
- preset/server: декомпозиция модуля с утилитами (patch)
- preset/node: утилита hideFirstId перенесена в preset/isomorphic (patch)
