+++ src/preset/server/index.ts
@@ -1,2 +1,5 @@
 export type { ServerHandler, ServerMiddleware, ServerHandlerContext } from './types';
-export { getClientIp, getForwardedHeaders, getPageResponseFormat, healthCheck } from './utils';
+export { getClientIp } from './utils/get-client-ip';
+export { getForwardedHeaders } from './utils/get-forwarded-headers';
+export { getHealthCheck } from './utils/get-health-check';
+export { getPageResponseFormat } from './utils/get-page-response-format';
