+++ src/preset/server/utils/get-page-response-format.ts
@@ -1,0 +1,16 @@
+/**
+ * Определяет формат ответа для страницы (html-верстки).
+ * Вернет "json" - если заголовок запроса "accept" содержит "application/json".
+ * Вернет "html" во всех остальных случаях.
+ * @param request Запрос.
+ * @return Формат.
+ */
+export function getPageResponseFormat(request: Request): 'html' | 'json' {
+  let result: 'html' | 'json' = 'html';
+
+  if ((request.headers.get('accept') || '').toLowerCase().includes('application/json')) {
+    result = 'json';
+  }
+
+  return result;
+}
